HRInterviewer/backend/requirements.txt
flask>=3.0
flask-cors>=4.0
numpy>=1.25
sounddevice>=0.5.3
pyttsx3>=2.90
pymongo>=4.6
groq>=1.0.0
google-genai>=1.56.0
python-dotenv>=1.0.1
elevenlabs>=2.26.0
faiss-cpu>=1.7.4
sentence-transformers>=2.7.0
pypdf>=5.1.0
